backend/requirements.txt
flask
python-dotenv
flask-cors
torch
torchvision

